test: organize HMR e2e test cases
