+++ e2e/cases/hmr/basic/rsbuild.config.ts
