+++ e2e/cases/hmr/basic/src/App.tsx
