+++ e2e/cases/hmr/basic/src/index.ts
