+++ e2e/cases/hmr/client-host/rsbuild.config.ts
