+++ e2e/cases/hmr/client-host/src/App.tsx
