+++ e2e/cases/hmr/client-host/src/index.ts
