+++ e2e/cases/hmr/client-port-placeholder/index.test.ts
@@ -1,0 +1,56 @@
+import fs from 'node:fs';
+import { join } from 'node:path';
+import { dev, getRandomPort, rspackOnlyTest } from '@e2e/helper';
+import { expect, test } from '@playwright/test';
+
+const cwd = __dirname;
+
+rspackOnlyTest(
+  'HMR should work when dev.client.port is `<port>`',
+  async ({ page }) => {
+    // HMR cases will fail on Windows
+    if (process.platform === 'win32') {
+      test.skip();
+    }
+
+    await fs.promises.cp(join(cwd, 'src'), join(cwd, 'test-temp-src'), {
+      recursive: true,
+    });
+
+    const port = await getRandomPort();
+    const rsbuild = await dev({
+      cwd,
+      page,
+      rsbuildConfig: {
+        source: {
+          entry: {
+            index: join(cwd, 'test-temp-src/index.ts'),
+          },
+        },
+        server: {
+          port,
+        },
+        dev: {
+          client: {
+            port: '<port>',
+          },
+        },
+      },
+    });
+
+    expect(rsbuild.port).toBe(port);
+
+    const appPath = join(cwd, 'test-temp-src/App.tsx');
+
+    const locator = page.locator('#test');
+    await expect(locator).toHaveText('Hello Rsbuild!');
+
+    await fs.promises.writeFile(
+      appPath,
+      fs.readFileSync(appPath, 'utf-8').replace('Hello Rsbuild', 'Hello Test'),
+    );
+
+    await expect(locator).toHaveText('Hello Test!');
+    await rsbuild.close();
+  },
+);
