+++ e2e/cases/hmr/client-port-placeholder/src/App.tsx
