+++ e2e/cases/hmr/error-recovery/index.test.ts
