+++ e2e/cases/hmr/error-recovery/src/App.css
@@ -1,0 +1,3 @@
+#test {
+  color: rgb(255, 0, 0);
+}
