+++ e2e/cases/hmr/error-recovery/src/App.tsx
@@ -1,0 +1,4 @@
+import './App.css';
+
+const App = () => <div id="test">Hello Rsbuild!</div>;
+export default App;
