+++ e2e/cases/hmr/rebuild-logs/index.test.ts
