+++ e2e/cases/hmr/rebuild-logs/src/App.tsx
@@ -1,0 +1,4 @@
+import './App.css';
+
+const App = () => <div id="test">Hello Rsbuild!</div>;
+export default App;
